// src/app/shared/Interfaces/boms_pieces_inspection.ts
export class BombPiecesInspection {
    work_order_id: number;
    order_detail_id: number;
    type_bomb_id: number;
    piece_bomb_id: number;
    type_piece: string;
    yes?: number;
    no?: number;
    repair?: number;
    supply?: number;
    demand?: number;
    stock?: number;
    description?: number;
}